Modified the file 3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -36,7 +36,7 @@ void print_all(const char * const format, ...)
 				printf("%d", i);
 				break;
 			case 'f':
-				f = (float) va_arg(args_list, double);
+				f =  va_arg(args_list, double);
 				printf("%f", f);
 				break;
 			case 's':
